requirements.txt
antlr4-python3-runtime==4.13.1
argcomplete==3.4.0
asttokens==2.4.1
attrs==23.2.0
black==24.4.2
click==8.1.7
cocotb==1.8.1
cocotb-AHB @ git+https://github.com/antmicro/cocotb-ahb@799e0de182fdc9a7ae7dc83c0ad4dfba353341e9
cocotb-bus==0.2.1
cocotb-coverage==1.2.0
cocotb-test==0.2.5
cocotbext-axi==0.1.24
-e git+https://github.com/alexforencich/cocotbext-i2c@3b32eedc15302f1299a23bb35aa669fffad2ca22#egg=cocotbext_i2c
git+https://github.com/antmicro/testplanner.git@bbf374a
-e ${I3C_ROOT_DIR}/third_party/cocotbext-i3c
-e ${I3C_ROOT_DIR}/tools/nox_utils
-e ${I3C_ROOT_DIR}/tools/cocotb_helpers
-e ${I3C_ROOT_DIR}/tools/vcd2pulseview
${I3C_ROOT_DIR}/tools/peakrdl_cocotb
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
engineering-notation==0.10.0
executing==2.0.1
filelock==3.15.4
find_libpython==0.4.0
flake8>=7
git-me-the-url==2.1.0
gitdb==4.0.11
GitPython==3.1.43
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.25.0
isort>=5
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nox==2023.4.22
numpy==2.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
peakrdl==1.1.0
peakrdl-cheader==1.0.0
peakrdl-html==2.10.1
peakrdl-ipxact==3.4.4
peakrdl-regblock==0.22.0
peakrdl-systemrdl==0.3.0
peakrdl-uvm==2.3.0
peakrdl_markdown==1.0.0
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.4.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-markdown-table==0.3.3
Pygments==2.18.0
pytest==8.1.1
pytest-html==4.1.1
pytest-md==0.2.0
pytest-metadata==3.1.1
pytest-timeout==2.3.1
python-constraint==1.4.0
python-dateutil==2.9.0.post0
python-markdown-math==0.8
pyuvm==2.9.1
pyyaml==6.0.1
pyzmq==26.0.3
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
systemrdl-compiler==1.27.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
virtualenv==20.26.3
wcwidth==0.2.13
zstandard==0.23.0
crc==7.0.0
